doc: use new pretty error
