+++ pretty/src/main/java/org/aya/pretty/error/PrettyError.java
@@ -7,9 +7,13 @@ import kala.collection.mutable.MutableList;
 import kala.control.Option;
 import kala.tuple.Tuple;
 import kala.tuple.Tuple2;
+import kala.tuple.Tuple3;
 import org.aya.pretty.doc.Doc;
 import org.aya.pretty.doc.Docile;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.Comparator;
 
 /**
  * @author kiva
@@ -92,6 +96,7 @@ public record PrettyError(
       Doc.plain("In file " + filePath + ":" + primary.startLine() + ":" + primary.startCol() + " ->"),
       Doc.empty(),
       Doc.hang(2, visualizeCode(config, allRange, primary, hints)),
+      Doc.empty(),
       brief,
       Doc.empty()
     );
@@ -104,6 +109,134 @@ public record PrettyError(
   private @NotNull String visualizeLine(@NotNull PrettyErrorConfig config, @NotNull String line) {
     int tabWidth = config.tabWidth();
     return line.replaceAll("\t", " ".repeat(tabWidth));
+  }
+
+  private static final class CodeBuilder {
+    private final @NotNull MutableList<Doc> lineDocs = MutableList.create();
+    private final @NotNull MutableList<Doc> codeDocs = MutableList.create();
+
+    void add(int currentLine, @NotNull Doc code) {
+      lineDocs.append(Doc.plain(String.valueOf(currentLine)));
+      codeDocs.append(code);
+    }
+
+    void add(@NotNull Doc code) {
+      lineDocs.append(Doc.plain(" ")); // cannot use `empty()` or `plain("")`
+      codeDocs.append(code);
+    }
+  }
+
+  private static final class HintGroup {
+    @NotNull MutableList<Doc> underlines = MutableList.create();
+    @NotNull MutableList<Doc> notes = MutableList.create();
+    @NotNull MutableList<HintLine> overlapped = MutableList.create();
+    @Nullable HintLine startOrEnd = null;
+
+    public void add(int indent, @NotNull HintLine line, @NotNull PrettyError.FormatConfig cfg) {
+      var hint = renderHint(cfg, line.startCol, line.endCol, switch (line.loc) {
+        case Shot, Between, None -> PrettyError.MultilineMode.DISABLED;
+        case Start -> PrettyError.MultilineMode.START;
+        case End -> PrettyError.MultilineMode.END;
+      });
+      underlines.append(Doc.indent(indent, hint));
+      notes.append(line.note);
+      if (line.loc == Span.NowLoc.Start || line.loc == Span.NowLoc.End)
+        startOrEnd = line;
+    }
+  }
+
+  record HintLine(Doc note, Span.NowLoc loc, int startCol, int endCol, int allocIndent) {}
+
+  private static final int INDENT_FACTOR = 2;
+
+  /** invariant: `others` is sorted by `startCol` */
+  private @NotNull HintGroup splitHints(@NotNull ImmutableSeq<HintLine> others) {
+    var group = new HintGroup();
+    if (others.isNotEmpty()) {
+      var first = others.first();
+      group.add(first.startCol, first, formatConfig);
+      int left = first.startCol;
+      int right = first.endCol;
+      int last = right;
+      for (int i = 1; i < others.size(); ++i) {
+        var the = others.get(i);
+        if (overlaps(left, right, the.startCol, the.endCol)) {
+          group.overlapped.append(the);
+        } else {
+          var indent = the.startCol - last;
+          group.add(indent, the, formatConfig);
+          last = the.endCol;
+          left = Math.min(left, the.startCol);
+          right = Math.max(right, the.endCol);
+        }
+      }
+    }
+    return group;
+  }
+
+  private void renderHints(
+    boolean continued,
+    int currentLine, int codeIndent,
+    int vbarUsedIndent, @NotNull Doc vbar,
+    @NotNull Doc currentCode,
+    @NotNull CodeBuilder builder,
+    @NotNull ImmutableSeq<HintLine> others
+  ) {
+    var split = splitHints(others);
+    var startOrEnd = split.startOrEnd;
+    var rest = codeIndent - vbarUsedIndent;
+    // commit code
+    if (!continued) {
+      var codeDoc = (startOrEnd == null || startOrEnd.loc != Span.NowLoc.End)
+        ? Doc.cat(vbar, Doc.indent(rest * INDENT_FACTOR, currentCode))
+        : Doc.cat(vbar, formatConfig.lineNoSepDoc(), Doc.indent(rest * INDENT_FACTOR - 1, currentCode));
+      builder.add(currentLine, codeDoc);
+    }
+    // commit hint
+    var almost = Doc.stickySep(Doc.cat(split.underlines), Doc.align(Doc.cat(split.notes)));
+    var codeHint = startOrEnd != null
+      ? renderStartEndHint(startOrEnd, vbar, almost, rest)
+      : renderContinuedHint(continued, vbar, almost, rest);
+    builder.add(codeHint);
+    // show overlapped hints in the next line
+    if (split.overlapped.isNotEmpty())
+      renderHints(true, currentLine, codeIndent, vbarUsedIndent, vbar, currentCode, builder, split.overlapped.toImmutableSeq());
+  }
+
+  private @NotNull Doc renderStartEndHint(@NotNull HintLine startOrEnd, @NotNull Doc vbar, @NotNull Doc almost, int rest) {
+    return startOrEnd.loc == Span.NowLoc.Start
+      ? Doc.cat(vbar, formatConfig.beginCornerDoc(), formatConfig.underlineBodyDoc(rest * INDENT_FACTOR - 1), almost)
+      : Doc.cat(vbar, formatConfig.endCornerDoc(), formatConfig.underlineBodyDoc(rest * INDENT_FACTOR - 1), almost);
+  }
+
+  private @NotNull Doc renderContinuedHint(boolean continued, @NotNull Doc vbar, @NotNull Doc almost, int rest) {
+    return continued && vbar.isEmpty()
+      ? Doc.cat(formatConfig.lineNoSepDoc(), Doc.indent(rest * INDENT_FACTOR, almost))
+      : Doc.cat(vbar, Doc.indent(rest * INDENT_FACTOR, almost));
+  }
+
+  private Tuple2<Integer, Doc> computeMultilineVBar(@NotNull ImmutableSeq<HintLine> between) {
+    int last = 0;
+    var vbar = Doc.empty();
+    for (var the : between) {
+      var level = the.allocIndent - last;
+      vbar = Doc.cat(vbar, Doc.indent(level * INDENT_FACTOR, formatConfig.lineNoSepDoc()));
+      last = the.endCol;
+    }
+    return Tuple.of(last, vbar);
+  }
+
+  private void renderHints(
+    int currentLine, int codeIndent,
+    @NotNull Doc currentCode,
+    @NotNull CodeBuilder builder,
+    @NotNull ImmutableSeq<HintLine> hintLines
+  ) {
+    var between = hintLines.filter(h -> h.loc == Span.NowLoc.Between)
+      .sorted(Comparator.comparingInt(a -> a.allocIndent));
+    var others = hintLines.filter(h -> h.loc != Span.NowLoc.Between);
+    var vbar = computeMultilineVBar(between);
+    renderHints(false, currentLine, codeIndent, vbar._1, vbar._2, currentCode, builder, others);
   }
 
   private @NotNull Doc visualizeCode(
@@ -125,90 +258,74 @@ public record PrettyError(
       .map(line -> visualizeLine(config, line))
       .collect(MutableList.factory());
 
-    int lineNo = Math.max(startLine - showMore, 1);
-    var docs = MutableList.<Doc>create();
+    final int minLineNo = Math.max(startLine - showMore, 1);
+    final int maxLineNo = minLineNo + lines.size();
+
+    // allocate hint nest
+    var alloc = MutableList.<Tuple3<Span.Data, Doc, Integer>>create();
+    alloc.append(Tuple.of(primaryRange, Doc.empty(), 0));
+    hints.view()
+      .filter(kv -> kv._1.startLine() >= minLineNo && kv._1.endLine() <= maxLineNo)
+      .mapIndexed((i, kv) -> Tuple.of(kv._1, kv._2, i + 1))
+      .forEach(alloc::append);
+
+    int codeIndent = alloc.size();
+    int lineNo = minLineNo;
+    var builder = new CodeBuilder();
 
     for (var line : lines) {
-      docs.append(renderLine(line, lineNo, linenoWidth));
+      final int currentLine = lineNo;
+      final var currentCode = Doc.plain(line);
 
-      // render error column as underlines
-      final int finalLineNo = lineNo;
-      hints.find(kv -> kv._1.startLine() == finalLineNo).ifDefined(find -> {
-        int startCol = find._1.startCol();
-        int endCol = find._1.endCol();
-        var hintUnderline = renderHint(startCol, endCol, linenoWidth, MultilineMode.DISABLED);
-        var doc = find._2;
-        if (doc instanceof Doc.Empty) docs.append(hintUnderline);
-        else docs.append(Doc.cat(hintUnderline, Doc.ONE_WS, doc));
-      });
-      // Inline hints are expected to be more informative. So we only
-      // underline primary error part if there's no inline hints.
-      if (hints.isEmpty()) {
-        if (primaryRange.startLine() == primaryRange.endLine()) {
-          if (lineNo == primaryRange.startLine()) {
-            var hintUnderline = renderHint(primaryRange.startCol(), primaryRange.endCol(), linenoWidth, MultilineMode.DISABLED);
-            docs.append(hintUnderline);
-          }
-        } else {
-          if (lineNo == primaryRange.startLine()) {
-            var hintUnderline = renderHint(primaryRange.startCol(), line.length(), linenoWidth, MultilineMode.START);
-            docs.append(Doc.stickySep(hintUnderline, Doc.align(Doc.english("Begin of the error"))));
-          } else if (lineNo == primaryRange.endLine()) {
-            var hintUnderline = renderHint(0, primaryRange.endCol(), linenoWidth, MultilineMode.END);
-            docs.append(Doc.stickySep(hintUnderline, Doc.align(Doc.english("End of the error"))));
-          }
-        }
+      var hintLines = alloc.view()
+        .mapNotNull(note -> switch (note._1.nowLoc(currentLine)) {
+          case Shot -> new HintLine(note._2, Span.NowLoc.Shot, note._1.startCol(), note._1.endCol(), note._3);
+          case Start -> new HintLine(Doc.empty(), Span.NowLoc.Start, 0, note._1.startCol(), note._3);
+          case End -> new HintLine(note._2, Span.NowLoc.End, 0, note._1.endCol(), note._3);
+          case Between -> new HintLine(Doc.empty(), Span.NowLoc.Between, 0, 0, note._3);
+          case None -> null;
+        })
+        .sorted(Comparator.comparingInt(a -> a.startCol))
+        .toImmutableSeq();
+
+      if (hintLines.isEmpty()) {
+        builder.add(currentLine, Doc.indent(codeIndent * INDENT_FACTOR, currentCode));
+      } else {
+        renderHints(currentLine, codeIndent, currentCode, builder, hintLines);
       }
       lineNo++;
     }
 
-    docs.append(Doc.empty());
-    return Doc.vcat(docs);
+    return Doc.catBlockR(linenoWidth,
+      builder.lineDocs,
+      Doc.cat(formatConfig.lineNoSepDoc(), Doc.ONE_WS),
+      builder.codeDocs);
   }
 
-  private Doc renderHint(int startCol, int endCol, int linenoWidth, MultilineMode mode) {
-    var builder = new StringBuilder();
-    builder.append(" ".repeat(linenoWidth));
-    if (formatConfig.vbarForHints().isDefined()) {
-      builder.append(' ');
-      builder.append(formatConfig.vbarForHints().get());
-      builder.append(" ".repeat(startCol + 1));
-    } else {
-      builder.append(" ".repeat(startCol + " | ".length()));
-    }
+  private static boolean overlaps(int x1, int x2, int y1, int y2) {
+    // assuming x1 <= x2, y1 <= y2
+    return x1 <= y2 && y1 <= x2;
+  }
 
-    if (mode == MultilineMode.DISABLED || formatConfig.errorIndicator().isEmpty()) {
-      builder.append(formatConfig.underlineBegin());
+  private static @NotNull Doc renderHint(@NotNull PrettyError.FormatConfig cfg, int startCol, int endCol, PrettyError.MultilineMode mode) {
+    var builder = new StringBuilder();
+    if (mode == PrettyError.MultilineMode.DISABLED || cfg.errorIndicator().isEmpty()) {
+      builder.append(cfg.underlineBegin());
       // -1 for the start symbol
       int length = endCol - startCol - 1;
       if (length > 0) {
         // endCol is in the next line
-        builder.append(Character.toString(formatConfig.underlineBody()).repeat(length));
+        builder.append(cfg.underlineBody(length));
       }
-      builder.append(formatConfig.underlineEnd());
+      builder.append(cfg.underlineEnd());
     } else {
       int length = endCol - startCol;
-      if (length > 0 && mode == MultilineMode.END) {
+      if (length > 0) {
         // endCol is in the next line
-        builder.append(" ".repeat(length));
+        builder.append(cfg.underlineBody(length));
       }
-      builder.append(formatConfig.errorIndicator().get());
+      builder.append(cfg.errorIndicator().get());
     }
-    return Doc.plain(builder.toString());
-  }
-
-  private Doc renderLine(String line, int lineNo, int linenoWidth) {
-    return renderLine(line, Option.ofNullable(lineNo), linenoWidth);
-  }
-
-  private Doc renderLine(String line, Option<Integer> lineNo, int linenoWidth) {
-    var builder = new StringBuilder();
-    if (lineNo.isDefined()) {
-      builder.append(String.format("%" + linenoWidth + "d %c ", lineNo.get(), formatConfig.lineNoSeparator()));
-    } else {
-      builder.append(String.format("%" + linenoWidth + "s %c ", "", formatConfig.lineNoSeparator()));
-    }
-    builder.append(line);
     return Doc.plain(builder.toString());
   }
